Separate out table component from page
