+++ hadoop-ozone/recon/src/main/resources/webapps/recon/ozone-recon-web/src/v2/pages/pipelines/pipelines.tsx
@@ -22,20 +22,13 @@ import React, {
   useState
 } from 'react';
 import moment from 'moment';
-import { Table, Tooltip } from 'antd';
-import { InfoCircleOutlined } from '@ant-design/icons';
-import {
-  ColumnsType,
-  TablePaginationConfig
-} from 'antd/es/table';
 import { ValueType } from 'react-select';
-import prettyMilliseconds from 'pretty-ms';
 
 import AutoReloadPanel from '@/components/autoReloadPanel/autoReloadPanel';
 import Search from '@/v2/components/search/search';
 import MultiSelect, { Option } from '@/v2/components/select/multiSelect';
+import PipelinesTable, { COLUMNS } from '@/v2/components/tables/pipelinesTable';
 import { showDataFetchError } from '@/utils/common';
-import { ReplicationIcon } from '@/utils/themeIcons';
 import { AutoReloadHelper } from '@/utils/autoReloadHelper';
 import { AxiosGetHelper, cancelRequests } from '@/utils/axiosRequestHelper';
 import { useDebounce } from '@/v2/hooks/debounce.hook';
@@ -43,152 +36,11 @@ import { useDebounce } from '@/v2/hooks/debounce.hook';
 import {
   Pipeline,
   PipelinesResponse,
-  PipelinesState,
-  PipelineStatusList
+  PipelinesState
 } from '@/v2/types/pipelines.types';
 
 import './pipelines.less';
-import { getDurationFromTimestamp, getTimeDiffFromTimestamp } from '@/v2/utils/momentUtils';
 
-// TODO: When Datanodes PR gets merged remove these declarations
-// And import from datanodes.types
-
-type SummaryDatanodeDetails = {
-  level: number;
-  parent: unknown | null;
-  cost: number;
-  uuid: string;
-  uuidString: string;
-  ipAddress: string;
-  hostName: string;
-  ports: {
-    name: string;
-    value: number
-  }[];
-  certSerialId: null,
-  version: string | null;
-  setupTime: number;
-  revision: string | null;
-  buildDate: string;
-  persistedOpState: string;
-  persistedOpStateExpiryEpochSec: number;
-  initialVersion: number;
-  currentVersion: number;
-  signature: number;
-  decommissioned: boolean;
-  networkName: string;
-  networkLocation: string;
-  networkFullPath: string;
-  numOfLeaves: number;
-}
-
-const COLUMNS: ColumnsType<Pipeline> = [
-  {
-    title: 'Pipeline ID',
-    dataIndex: 'pipelineId',
-    key: 'pipelineId',
-    sorter: (a: Pipeline, b: Pipeline) => a.pipelineId.localeCompare(b.pipelineId),
-
-  },
-  {
-    title: 'Replication Type & Factor',
-    dataIndex: 'replicationType',
-    key: 'replicationType',
-    render: (replicationType: string, record: Pipeline) => {
-      const replicationFactor = record.replicationFactor;
-      return (
-        <span>
-          <ReplicationIcon
-            replicationFactor={replicationFactor}
-            replicationType={replicationType}
-            leaderNode={record.leaderNode}
-            isLeader={false} />
-          {replicationType} ({replicationFactor})
-        </span>
-      );
-    },
-    sorter: (a: Pipeline, b: Pipeline) =>
-      (a.replicationType + a.replicationFactor.toString()).localeCompare(b.replicationType + b.replicationFactor.toString()),
-    defaultSortOrder: 'descend' as const
-  },
-  {
-    title: 'Status',
-    dataIndex: 'status',
-    key: 'status',
-    filterMultiple: true,
-    filters: PipelineStatusList.map(status => ({ text: status, value: status })),
-    onFilter: (value, record: Pipeline) => record.status === value,
-    sorter: (a: Pipeline, b: Pipeline) => a.status.localeCompare(b.status)
-  },
-  {
-    title: 'Containers',
-    dataIndex: 'containers',
-    key: 'containers',
-    sorter: (a: Pipeline, b: Pipeline) => a.containers - b.containers
-  },
-  {
-    title: 'Datanodes',
-    dataIndex: 'datanodes',
-    key: 'datanodes',
-    render: (datanodes: SummaryDatanodeDetails[]) => (
-      <div>
-        {datanodes.map(datanode => (
-          <div className='uuid-tooltip'>
-            <Tooltip
-              placement='top'
-              title={`UUID: ${datanode?.uuid ?? 'NA'}`}
-              getPopupContainer={(triggerNode) => triggerNode}>
-              {datanode?.hostName ?? 'N/A'}
-            </Tooltip>
-          </div>
-        ))}
-      </div>
-    )
-  },
-  {
-    title: 'Leader',
-    dataIndex: 'leaderNode',
-    key: 'leaderNode',
-    sorter: (a: Pipeline, b: Pipeline) => a.leaderNode.localeCompare(b.leaderNode)
-  },
-  {
-    title: () => (
-      <span>
-        Last Leader Election&nbsp;
-        <Tooltip title='Elapsed time since the current leader got elected. Only available if any metrics service providers like Prometheus is configured.'>
-          <InfoCircleOutlined />
-        </Tooltip>
-      </span>
-    ),
-    dataIndex: 'lastLeaderElection',
-    key: 'lastLeaderElection',
-    render: (lastLeaderElection: number) => lastLeaderElection > 0 ?
-      getTimeDiffFromTimestamp(lastLeaderElection) : 'NA',
-    sorter: (a: Pipeline, b: Pipeline) => a.lastLeaderElection - b.lastLeaderElection
-  },
-  {
-    title: 'Lifetime',
-    dataIndex: 'duration',
-    key: 'duration',
-    render: (duration: number) => getDurationFromTimestamp(duration),
-    sorter: (a: Pipeline, b: Pipeline) => a.duration - b.duration
-  },
-  {
-    title: () => (
-      <span>
-        No. of Elections&nbsp;
-        <Tooltip title='Number of elections in this pipeline. Only available if any metrics service providers like Prometheus is configured.'>
-          <InfoCircleOutlined />
-        </Tooltip>
-      </span>
-    ),
-    dataIndex: 'leaderElections',
-    key: 'leaderElections',
-    render: (leaderElections: number) => leaderElections > 0 ?
-      leaderElections : 'NA',
-    sorter: (a: Pipeline, b: Pipeline) => a.leaderElections - b.leaderElections
-  }
-];
 
 const defaultColumns = COLUMNS.map(column => ({
   label: (typeof column.title === 'string')
@@ -248,19 +100,6 @@ const Pipelines: React.FC<{}> = () => {
     })
   }, []);
 
-  function filterSelectedColumns() {
-    const columnKeys = selectedColumns.map((column) => column.value);
-    return COLUMNS.filter(
-      (column) => columnKeys.indexOf(column.key as string) >= 0
-    )
-  }
-
-  function getFilteredData(data: Pipeline[]) {
-    return data.filter(
-      (pipeline: Pipeline) => pipeline['pipelineId'].includes(debouncedSearch)
-    )
-  }
-
   function handleColumnChange(selected: ValueType<Option, true>) {
     setSelectedColumns(selected as Option[]);
   }
@@ -270,10 +109,7 @@ const Pipelines: React.FC<{}> = () => {
     columnOptions,
     lastUpdated
   } = state;
-  const paginationConfig: TablePaginationConfig = {
-    showTotal: (total: number, range) => `${range[0]}-${range[1]} of ${total} pipelines`,
-    showSizeChanger: true,
-  };
+
   return (
     <>
       <div className='page-header-v2'>
@@ -311,16 +147,11 @@ const Pipelines: React.FC<{}> = () => {
               }
               onChange={() => { }} />
           </div>
-          <div>
-            <Table
-              dataSource={getFilteredData(activeDataSource)}
-              columns={filterSelectedColumns()}
-              loading={loading}
-              rowKey='pipelineId'
-              pagination={paginationConfig}
-              scroll={{ x: 'max-content', scrollToFirstRowOnChange: true }}
-              locale={{ filterTitle: '' }} />
-          </div>
+          <PipelinesTable
+            loading={loading}
+            data={activeDataSource}
+            selectedColumns={selectedColumns}
+            searchTerm={debouncedSearch} />
         </div>
       </div>
     </>
